crewai/crewai-project/pyproject.toml
[project]
name = "crewai_project"
version = "0.1.0"
description = "Default template for PDM package"
authors = [
    {name = "martimfasantos", email = "martimfasantos@gmail.com"},
]
dependencies = ["starlette>=0.45.2", "pydantic-core>=2.27.2", "annotated-types>=0.7.0", "crewai-tools>=0.0.1", "crewai[tools]>=0.100.1", "pydantic-settings>=2.7.1"]
requires-python = "<3.13,>=3.12"
readme = "README.md"
license = {text = "MIT"}

[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"


[tool.pdm]
distribution = true

[tool.pdm.scripts]
crewai = "python3 src/crewai_project/main.py"
